Use GNU Readline if it's available
A couple of popular Gems (Pry, Guard) do not function properly without
using GNU readline (e.g. [this GitHub
Issue](https://github.com/guard/guard/wiki/Add-Readline-support-to-Ruby-on-Mac-OS-X)).

This commit checks to see if GNU Readline has been installed via
homebrew on the Mac OS X and adds the proper `configure` argument. I
don't *think* this is an issue on Linux because GNU Readline seems to be
the default there.